rebpf: separate the BPF and userspace map traits
It turns out that there are some notions about those traits that are
useful for userspace but not for BPF. The first step is thus to actually
have separate traits for both subsystems.
